исправил конфликты

+++ src/main/java/ru/skypro/homework/dto/comment/CommentDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CommentDTO {
-    private int author;
+    private Long author;
     private String authorImage;
     private String authorFirstName;
     private int createdAt;
